--- web/apps/photos/src/components/Sidebar/Preferences/LanguageSelector.tsx
@@ -1,61 +1,0 @@
-import {
-    getLocaleInUse,
-    setLocaleInUse,
-    supportedLocales,
-    type SupportedLocale,
-} from "@/next/i18n";
-import DropdownInput, { DropdownOption } from "components/DropdownInput";
-import { t } from "i18next";
-import { useRouter } from "next/router";
-
-/**
- * Human readable name for each supported locale.
- */
-export const localeName = (locale: SupportedLocale) => {
-    switch (locale) {
-        case "en-US":
-            return "English";
-        case "fr-FR":
-            return "Français";
-        case "de-DE":
-            return "Deutsch";
-        case "zh-CN":
-            return "中文";
-        case "nl-NL":
-            return "Nederlands";
-        case "es-ES":
-            return "Español";
-        case "pt-BR":
-            return "Brazilian Portuguese";
-        case "ru-RU":
-            return "Russian";
-    }
-};
-
-const getLanguageOptions = (): DropdownOption<SupportedLocale>[] => {
-    return supportedLocales.map((locale) => ({
-        label: localeName(locale),
-        value: locale,
-    }));
-};
-
-export const LanguageSelector = () => {
-    const locale = getLocaleInUse();
-    // Enhancement: Is this full reload needed?
-    const router = useRouter();
-
-    const updateCurrentLocale = (newLocale: SupportedLocale) => {
-        setLocaleInUse(newLocale);
-        router.reload();
-    };
-
-    return (
-        <DropdownInput
-            options={getLanguageOptions()}
-            label={t("LANGUAGE")}
-            labelProps={{ color: "text.muted" }}
-            selected={locale}
-            setSelected={updateCurrentLocale}
-        />
-    );
-};
